Move map’s controls in controls folder
