+++ components/map/controls/control.js
