+++ components/map/map.js
@@ -12,14 +12,14 @@ import DrawContext from '@/contexts/draw'
 import ParcellesContext from '@/contexts/parcelles'
 
 import {vector, ortho, planIGN} from '@/components/map/styles'
-import NavControl from '@/components/map/nav-control'
 import EditableMarker from '@/components/map/editable-marker'
-import Control from '@/components/map/control'
 import NumerosMarkers from '@/components/map/numeros-markers'
 import ToponymeMarker from '@/components/map/toponyme-marker'
 import Draw from '@/components/map/draw'
-import StyleSelector from '@/components/map/style-selector'
-import AddressEditorControl from '@/components/map/address-editor-control'
+import Control from '@/components/map/controls/control'
+import NavControl from '@/components/map/controls/nav-control'
+import StyleSelector from '@/components/map/controls/style-selector'
+import AddressEditorControl from '@/components/map/controls/address-editor-control'
 import useBounds from '@/components/map/hooks/bounds'
 import useSources from '@/components/map/hooks/sources'
 import useLayers from '@/components/map/hooks/layers'
